+++ src/hypercall/mod.rs
@@ -1,7 +1,7 @@
 use core::cmp;
 use std::{
 	collections::BTreeMap,
-	ffi::{CStr, CString, OsStr},
+	ffi::{CStr, CString},
 	fs, io,
 	os::{fd::IntoRawFd, unix::ffi::OsStrExt},
 	path::Path,
@@ -11,8 +11,7 @@ use std::{
 use fstat::{fstat, stat};
 use getdents::getdents;
 use uhyve_interface::{
-	GuestPhysAddr,
-	v1::{self, MAX_ARGC_ENVC},
+	GuestPhysAddr, v1,
 	v2::{
 		self,
 		parameters::{FileType, *},
@@ -28,12 +27,12 @@ use crate::{
 	mem::MmapMemory,
 	mem_layout::MemoryLayout,
 	net::NetworkBackend,
-	params::EnvVars,
 	vcpu::VcpuStopReason,
 	virt_to_phys,
 	vm::{KernelInfo, VmPeripherals},
 };
 
+mod cmdval_copy;
 mod fstat;
 mod getdents;
 
@@ -201,13 +200,13 @@ pub fn handle_hypercall_v1<N: NetworkBackend, L: MemoryLayout>(
 			syssize.update(&kernel_info.path, &kernel_info.params.kernel_args)
 		}
 		v1::Hypercall::Cmdval(syscmdval) => {
-			copy_argv(
+			cmdval_copy::copy_argv(
 				kernel_info.path.as_os_str(),
 				&kernel_info.params.kernel_args,
 				syscmdval,
 				&peripherals.mem,
 			);
-			copy_env(&kernel_info.params.env, syscmdval, &peripherals.mem);
+			cmdval_copy::copy_env(&kernel_info.params.env, syscmdval, &peripherals.mem);
 		}
 		v1::Hypercall::Exit(sysexit) => {
 			return Some(Ok(VcpuStopReason::Exit(sysexit.arg)));
@@ -775,77 +774,4 @@ pub(crate) fn mkdir(mem: &MmapMemory, sysmkdir: &mut MkdirParams, file_map: &mut
 			}
 		}
 	};
-}
-
-/// Copies the arguments of the application into the VM's memory to the destinations specified in `syscmdval`.
-fn copy_argv(
-	path: &OsStr,
-	argv: &[String],
-	syscmdval: &v1::parameters::CmdvalParams,
-	mem: &MmapMemory,
-) {
-	// copy kernel path as first argument
-	let argvp = mem
-		.host_address(syscmdval.argv)
-		.expect("Systemcall parameters for Cmdval are invalid") as *const GuestPhysAddr;
-	let arg_addrs = unsafe { std::slice::from_raw_parts(argvp, argv.len() + 1) };
-
-	{
-		let len = path.len();
-		// Safety: we drop path_dest before anything else is done with mem
-		let path_dest = unsafe {
-			mem.slice_at_mut(arg_addrs[0], len + 1)
-				.expect("Systemcall parameters for Cmdval are invalid")
-		};
-
-		path_dest[0..len].copy_from_slice(path.as_bytes());
-		path_dest[len] = 0; // argv strings are zero terminated
-	}
-
-	// Copy the application arguments into the vm memory
-	for (argument, &arg_dest_addr) in argv.iter().zip(arg_addrs.iter()) {
-		let len = argument.len();
-		let arg_dest = unsafe {
-			mem.slice_at_mut(arg_dest_addr, len + 1)
-				.expect("Systemcall parameters for Cmdval are invalid")
-		};
-		arg_dest[0..len].copy_from_slice(argument.as_bytes());
-		arg_dest[len] = 0;
-	}
-}
-
-/// Copies the environment variables into the VM's memory to the destinations specified in `syscmdval`.
-fn copy_env(env: &EnvVars, syscmdval: &v1::parameters::CmdvalParams, mem: &MmapMemory) {
-	let envp = mem
-		.host_address(syscmdval.envp)
-		.expect("Systemcall parameters for Cmdval are invalid") as *const GuestPhysAddr;
-
-	let env: Vec<(String, String)> = match env {
-		EnvVars::Host => std::env::vars().collect(),
-		EnvVars::Set(map) => map
-			.iter()
-			.map(|(a, b)| (a.to_owned(), b.to_owned()))
-			.collect(),
-	};
-	if env.len() >= MAX_ARGC_ENVC {
-		warn!(
-			"Environment is larger than the maximum that can be copied to the VM. Remaining environment is ignored"
-		);
-	}
-	let env_addrs = unsafe { std::slice::from_raw_parts(envp, env.len()) };
-
-	// Copy the environment variables into the vm memory
-	for ((key, value), &env_dest_addr) in env.iter().take(MAX_ARGC_ENVC).zip(env_addrs.iter()) {
-		let len = key.len() + value.len() + 1;
-		let env_dest = unsafe {
-			mem.slice_at_mut(env_dest_addr, len + 1)
-				.expect("Systemcall parameters for Cmdval are invalid")
-		};
-		//write_env_into_mem(env_dest, key.as_bytes(), value.as_bytes());
-		let len = key.len() + value.len() + 1;
-		env_dest[0..key.len()].copy_from_slice(key.as_bytes());
-		env_dest[key.len()] = b'=';
-		env_dest[key.len() + 1..len].copy_from_slice(value.as_bytes());
-		env_dest[len] = 0;
-	}
 }
